[Module-rename] update rest of files with new module names
