+++ fnl/katdotnvim/color.fnl
@@ -1,5 +1,5 @@
 (module katdotnvim.color
-        {autoload {ucolors katdotnvim.utils.color main katdotnvim.main}})
+        {autoload {ucolors katdotnvim.utils.highlight.utils main katdotnvim.main}})
 
 ;;; This stores and exports all needed colors as functions
