+++ fnl/katdotnvim/highlights/integrations/indent_blankline.fnl
@@ -1,5 +1,5 @@
 (module katdotnvim.highlights.integrations.indent_blankline
-        {autoload {ucolors katdotnvim.utils.color
+        {autoload {ucolors katdotnvim.utils.highlight.utils
                    colors katdotnvim.color
                    syntax katdotnvim.highlights.syntax
                    run katdotnvim.utils.highlight.run
