+++ fnl/katdotnvim/highlights/integrations/lsp.fnl
@@ -1,5 +1,5 @@
 (module katdotnvim.highlights.integrations.lsp
-        {autoload {ucolors katdotnvim.utils.color
+        {autoload {ucolors katdotnvim.utils.highlight.utils
                    colors katdotnvim.color
                    syntax katdotnvim.highlights.syntax
                    run katdotnvim.utils.highlight.run
