+++ fnl/katdotnvim/highlights/main.fnl
@@ -1,6 +1,6 @@
 (module katdotnvim.highlights.main
         {autoload {colors katdotnvim.color
-                   ucolors katdotnvim.utils.color
+                   ucolors katdotnvim.utils.highlight.utils
                    run katdotnvim.utils.highlight.run
                    main katdotnvim.main}})
