+++ fnl/katdotnvim/highlights/syntax.fnl
@@ -1,6 +1,6 @@
 (module katdotnvim.highlights.syntax
         {autoload {colors katdotnvim.color
-                   ucolors katdotnvim.utils.color
+                   ucolors katdotnvim.utils.highlight.utils
                    groups katdotnvim.highlights.main
                    run katdotnvim.utils.highlight.run
                    main katdotnvim.main}})
