+++ fnl/katdotnvim/stupid.fnl
@@ -1,4 +1,5 @@
-(module katdotnvim.stupid {autoload {ucolors katdotnvim.utils.color
+(module katdotnvim.stupid {autoload {ucolors katdotnvim.utils.highlight.utils
+                                     apply katdotnvim.utils.highlight.init
                                      colors katdotnvim.color
                                      syntax katdotnvim.highlights.syntax
                                      a aniseed.core}})
@@ -25,7 +26,7 @@
                                        0.1)
                                      0.5)
                                :bold true}]
-                    (ucolors.highlight$ opts)
+                    (apply.highlight$ opts)
                     (if (= increase true)
                       (set i (+ i 1))
                       (set i (- i 1)))
