+++ fnl/katdotnvim/utils/export/kitty.fnl
@@ -1,6 +1,6 @@
 (module katdotnvim.utils.export.kitty
         {autoload {hsl externals.hsluv
-                   ucolors katdotnvim.utils.color
+                   ucolors katdotnvim.utils.highlight.utils
                    colors katdotnvim.color
                    groups katdotnvim.highlights.main
                    main katdotnvim.main
