+++ fnl/katdotnvim/utils/export/konsole.fnl
@@ -1,6 +1,6 @@
 (module katdotnvim.utils.export.konsole
         {autoload {hsl externals.hsluv
-                   ucolors katdotnvim.utils.color
+                   ucolors katdotnvim.utils.highlight.utils
                    colors katdotnvim.color
                    groups katdotnvim.highlights.main
                    main katdotnvim.main
