+++ fnl/katdotnvim/utils/highlight/run.fnl
@@ -1,5 +1,6 @@
 (module katdotnvim.utils.highlight.run
-        {autoload {a aniseed.core ucolors katdotnvim.utils.color}})
+        {autoload {a aniseed.core ucolors katdotnvim.utils.highlight.utils
+                   apply katdotnvim.utils.highlight.init}})
 
 ;;; Handles running of highlight groups
 
@@ -15,12 +16,12 @@ Table is: group guifg guibg termfg termbg attr special blend"
         (if (and (= (type value) :function) (not= (value) nil))
             (do
               (each [_ nest (pairs [(value)])]
-                (ucolors.highlight$ nest)))
+                (apply.highlight$ nest)))
             ; everything else should just be a table, but still make sure that ; the nil returning function doesn't go through
             (or (= (type value) :table) (not= (value) nil))
             (do
               ; just check if we have a nested table or not
               (if (= (type (?. value 1)) :table)
                   (each [_ nest (pairs value)]
-                    (ucolors.highlight$ nest))
-                  (ucolors.highlight$ value))))))
+                    (apply.highlight$ nest))
+                  (apply.highlight$ value))))))
