+++ fnl/lualine/themes/kat.fnl
@@ -1,4 +1,4 @@
-(local ucolors (require :katdotnvim.utils.color))
+(local ucolors (require :katdotnvim.utils.highlight.utils))
 (local colors (require :katdotnvim.color))
 (local syntax (require :katdotnvim.highlights.syntax))
 (local groups (require :katdotnvim.highlights.main))
