+++ lua/katdotnvim/main.lua
@@ -11,10 +11,9 @@ do
   _2amodule_locals_2a = (_2amodule_2a)["aniseed/locals"]
 end
 local autoload = (require("aniseed.autoload")).autoload
-local colors, options, ucolors, _ = autoload("katdotnvim.color"), autoload("katdotnvim.utils.options.init"), autoload("katdotnvim.utils.color"), nil
+local colors, options, _ = autoload("katdotnvim.color"), autoload("katdotnvim.utils.options.init"), nil
 _2amodule_locals_2a["colors"] = colors
 _2amodule_locals_2a["options"] = options
-_2amodule_locals_2a["ucolors"] = ucolors
 _2amodule_locals_2a["_"] = _
 local function init(in_contrast)
   options.default(true)
