+++ lua/katdotnvim/utils/export/kitty.lua
@@ -11,7 +11,7 @@ do
   _2amodule_locals_2a = (_2amodule_2a)["aniseed/locals"]
 end
 local autoload = (require("aniseed.autoload")).autoload
-local a, colors, errors, export, groups, hsl, main, ucolors = autoload("aniseed.core"), autoload("katdotnvim.color"), autoload("katdotnvim.utils.errors"), autoload("katdotnvim.utils.export.init"), autoload("katdotnvim.highlights.main"), autoload("externals.hsluv"), autoload("katdotnvim.main"), autoload("katdotnvim.utils.color")
+local a, colors, errors, export, groups, hsl, main, ucolors = autoload("aniseed.core"), autoload("katdotnvim.color"), autoload("katdotnvim.utils.errors"), autoload("katdotnvim.utils.export.init"), autoload("katdotnvim.highlights.main"), autoload("externals.hsluv"), autoload("katdotnvim.main"), autoload("katdotnvim.utils.highlight.utils")
 do end (_2amodule_locals_2a)["a"] = a
 _2amodule_locals_2a["colors"] = colors
 _2amodule_locals_2a["errors"] = errors
