+++ lua/katdotnvim/utils/highlight/init.lua
@@ -11,110 +11,116 @@ do
   _2amodule_locals_2a = (_2amodule_2a)["aniseed/locals"]
 end
 local autoload = (require("aniseed.autoload")).autoload
-local a, s = autoload("aniseed.core"), autoload("aniseed.string")
-do end (_2amodule_locals_2a)["a"] = a
+local a, colors, get, hsl, s, _ = autoload("aniseed.core"), autoload("katdotnvim.color"), autoload("katdotnvim.utils.highlight.get"), autoload("externals.hsluv"), autoload("aniseed.string"), nil
+_2amodule_locals_2a["a"] = a
+_2amodule_locals_2a["colors"] = colors
+_2amodule_locals_2a["get"] = get
+_2amodule_locals_2a["hsl"] = hsl
 _2amodule_locals_2a["s"] = s
-local function group(table)
-  local t_1_ = table
-  if (nil ~= t_1_) then
-    t_1_ = (t_1_).group
+_2amodule_locals_2a["_"] = _
+local function highlight_24_3c_vimscript(opts)
+  if get.link(opts) then
+    local group = get.group(opts)
+    local link = get.link(opts)
+    return vim.cmd(string.format("hi def link %s %s", group, link))
   else
-  end
-  return t_1_
-end
-_2amodule_2a["group"] = group
-local function gui_fg(table)
-  local t_3_ = table
-  if (nil ~= t_3_) then
-    t_3_ = (t_3_).fg
-  else
-  end
-  return t_3_
-end
-_2amodule_2a["gui-fg"] = gui_fg
-local function gui_bg(table)
-  local t_5_ = table
-  if (nil ~= t_5_) then
-    t_5_ = (t_5_).bg
-  else
-  end
-  return t_5_
-end
-_2amodule_2a["gui-bg"] = gui_bg
-local function term_fg(table)
-  local t_7_ = table
-  if (nil ~= t_7_) then
-    t_7_ = (t_7_).ctermfg
-  else
-  end
-  return t_7_
-end
-_2amodule_2a["term-fg"] = term_fg
-local function term_bg(table)
-  local t_9_ = table
-  if (nil ~= t_9_) then
-    t_9_ = (t_9_).ctermbg
-  else
-  end
-  return t_9_
-end
-_2amodule_2a["term-bg"] = term_bg
-local function special(table)
-  local t_11_ = table
-  if (nil ~= t_11_) then
-    t_11_ = (t_11_).sp
-  else
-  end
-  return t_11_
-end
-_2amodule_2a["special"] = special
-local function blend(table)
-  local t_13_ = table
-  if (nil ~= t_13_) then
-    t_13_ = (t_13_).blend
-  else
-  end
-  return t_13_
-end
-_2amodule_2a["blend"] = blend
-local function link(table)
-  local t_15_ = table
-  if (nil ~= t_15_) then
-    t_15_ = (t_15_).link
-  else
-  end
-  return t_15_
-end
-_2amodule_2a["link"] = link
-local function attr__3etable(table_23)
-  local output = {}
-  for k, v in pairs(table_23) do
-    if (v == true) then
-      table.insert(output, k)
+    local group = get.group(opts)
+    local gui_fore
+    if ((get["gui-fg"](opts) ~= nil) and (opts.fg ~= "SKIP")) then
+      gui_fore = string.format(" guifg=%s", opts.fg)
     else
+      gui_fore = ""
     end
-  end
-  return output
-end
-_2amodule_2a["attr->table"] = attr__3etable
-local function all_attr__3etable(table_23)
-  local output = {}
-  for k, v in pairs(table_23) do
-    if ((v == true) or (v == false)) then
-      output[k] = v
+    local gui_back
+    if ((get["gui-bg"](opts) ~= nil) and (opts.bg ~= "SKIP")) then
+      gui_back = string.format(" guibg=%s", opts.bg)
     else
+      gui_back = ""
     end
+    local c_fore
+    if ((get["term-fg"](opts) ~= nil) and (opts.ctermfg ~= "SKIP")) then
+      c_fore = string.format(" ctermfg=%s", opts.ctermfg)
+    else
+      c_fore = ""
+    end
+    local c_back
+    if ((get["term-bg"](opts) ~= nil) and (opts.ctermbg ~= "SKIP")) then
+      c_back = string.format(" ctermbg=%s", opts.ctermbg)
+    else
+      c_back = ""
+    end
+    local special
+    if (get.special(opts) ~= nil) then
+      special = string.format(" guisp=%s", get.special(opts))
+    else
+      special = ""
+    end
+    local blend
+    if (get.blend(opts) ~= nil) then
+      blend = string.format(" blend=%s", get.blend(opts))
+    else
+      blend = ""
+    end
+    local attr
+    do
+      local attr_string = get["attr->string"](opts)
+      if (attr_string ~= nil) then
+        attr = string.format(" gui=%s cterm=%s", attr_string, attr_string)
+      else
+        attr = ""
+      end
+    end
+    local highlight = ("highlight " .. group .. gui_fore .. gui_back .. c_fore .. c_back .. attr .. special .. blend)
+    return vim.cmd(tostring(highlight))
   end
-  return output
 end
-_2amodule_2a["all-attr->table"] = all_attr__3etable
-local function attr__3estring(table)
-  local attr_table = attr__3etable(table)
-  if a["empty?"](attr_table) then
-    return nil
+_2amodule_locals_2a["highlight$<-vimscript"] = highlight_24_3c_vimscript
+local function highlight_24_3c_api(opts)
+  if get.link(opts) then
+    local group = get.group(opts)
+    local link = get.link(opts)
+    local args = {link = link}
+    return vim.api.nvim_set_hl(0, group, args)
   else
-    return s.join(",", attr__3etable(table))
+    local group = get.group(opts)
+    local gui_fore
+    if ((get["gui-fg"](opts) ~= nil) and (opts.fg ~= "NONE") and (opts.fg ~= "SKIP")) then
+      gui_fore = opts.fg
+    else
+      gui_fore = nil
+    end
+    local gui_back
+    if ((get["gui-bg"](opts) ~= nil) and (opts.bg ~= "NONE") and (opts.bg ~= "SKIP")) then
+      gui_back = opts.bg
+    else
+      gui_back = nil
+    end
+    local c_fore
+    if ((get["term-fg"](opts) ~= nil) and (opts.ctermfg ~= "NONE") and (opts.ctermfg ~= "SKIP")) then
+      c_fore = opts.ctermfg
+    else
+      c_fore = nil
+    end
+    local c_back
+    if ((get["term-bg"](opts) ~= nil) and (opts.ctermbg ~= "NONE") and (opts.ctermbg ~= "SKIP")) then
+      c_back = opts.ctermbg
+    else
+      c_back = nil
+    end
+    local args = {fg = gui_fore, bg = gui_back, ctermfg = c_fore, ctermbg = c_back, special = get.special(opts), blend = get.blend(opts)}
+    for k, v in pairs(get["all-attr->table"](opts)) do
+      args[k] = v
+    end
+    return vim.api.nvim_set_hl(0, group, args)
   end
 end
-_2amodule_2a["attr->string"] = attr__3estring
+_2amodule_locals_2a["highlight$<-api"] = highlight_24_3c_api
+local function highlight_24(opts)
+  if (vim.fn.has("nvim-0.7") == 0) then
+    return highlight_24_3c_vimscript(opts)
+  else
+    return highlight_24_3c_api(opts)
+  end
+end
+_2amodule_2a["highlight$"] = highlight_24
 return _2amodule_2a
