+++ lua/katdotnvim/utils/highlight/run.lua
@@ -11,14 +11,15 @@ do
   _2amodule_locals_2a = (_2amodule_2a)["aniseed/locals"]
 end
 local autoload = (require("aniseed.autoload")).autoload
-local a, ucolors = autoload("aniseed.core"), autoload("katdotnvim.utils.color")
+local a, apply, ucolors = autoload("aniseed.core"), autoload("katdotnvim.utils.highlight.init"), autoload("katdotnvim.utils.highlight.utils")
 do end (_2amodule_locals_2a)["a"] = a
+_2amodule_locals_2a["apply"] = apply
 _2amodule_locals_2a["ucolors"] = ucolors
 local function highlight_24_3c_table(high_table)
   for _, value in pairs(high_table) do
     if ((type(value) == "function") and (value() ~= nil)) then
       for _0, nest in pairs({value()}) do
-        ucolors["highlight$"](nest)
+        apply["highlight$"](nest)
       end
     elseif ((type(value) == "table") or (value() ~= nil)) then
       local function _2_()
@@ -31,10 +32,10 @@ local function highlight_24_3c_table(high_table)
       end
       if (type(_2_()) == "table") then
         for _0, nest in pairs(value) do
-          ucolors["highlight$"](nest)
+          apply["highlight$"](nest)
         end
       else
-        ucolors["highlight$"](value)
+        apply["highlight$"](value)
       end
     else
     end
